added validation and error handling

+++ controllers/contacts.js
@@ -1,7 +1,7 @@
 const mongodb = require('../data/database'); //connection
-const Objectid = require('mongodb').ObjectId;
+const ObjectId = require('mongodb').ObjectId;
 
-const getAll = async (req, res) => {
+/*const getAll = async (req, res) => {
     //#swagger.tags=['Contacts']
     const result = await mongodb.getDatabase().db().collection('contacts').find();
     result.toArray().then((contacts) => {
@@ -9,9 +9,36 @@ const getAll = async (req, res) => {
         res.status(200).json(contacts);
 
     });
+};*/
+
+/*const getAll = (req, res) => {
+    //#swagger.tags=['Contacts']
+    mongodb
+        .getDatabase()
+        .db()
+        .collection('contacts')
+        .find()
+        .toArray((err, lists) => {
+            if (err) {
+                res.status(400).json({ message: err });
+            }
+            res.setHeader('Content-Type', 'application/json');
+            res.status(200).json(lists);
+        });
+};*/
+
+const getAll = async (req, res) => {
+    //#swagger.tags=['Contacts']
+    try {
+        const result = await mongodb.getDatabase().db().collection('contacts').find().toArray();
+        res.setHeader('Content-Type', 'application/json');
+        res.status(200).json(result);
+    } catch (err) {
+        res.status(500).json({ message: err.message });
+    }
 };
 
-const getSingle = async (req, res) => {
+/*const getSingle = async (req, res) => {
     //#swagger.tags=['Contacts']
     try {
         const contactId = new Objectid(req.params.id);
@@ -27,9 +54,48 @@ const getSingle = async (req, res) => {
     } catch (error) {
         res.status(400).json({ message: 'Invalid contact id format' });
     }
+};*/
+
+/*const getSingle = (req, res) => {
+    //#swagger.tags=['Contacts']
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json('Must use a valid contact id to find a contact.');
+    }
+    const userId = new ObjectId(req.params.id);
+    mongodb
+        .getDatabase()
+        .db()
+        .collection('contacts')
+        .find({ _id: userId })
+        .toArray((err, result) => {
+            if (err) {
+                res.status(400).json({ message: err });
+            }
+            res.setHeader('Content-Type', 'application/json');
+            res.status(200).json(result[0]);
+        });
+};*/
+
+const getSingle = async (req, res) => {
+    //#swagger.tags=['Contacts']
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json('Must use a valid contact id to find a contact.');
+    }
+    try {
+        const userId = new ObjectId(req.params.id);
+        const result = await mongodb.getDatabase().db().collection('contacts').find({ _id: userId }).toArray();
+        if (!result[0]) {
+            return res.status(404).json({ message: 'Contact not found' });
+        }
+        res.setHeader('Content-Type', 'application/json');
+        res.status(200).json(result[0]);
+    } catch (err) {
+        res.status(500).json({ message: err.message });
+    }
 };
 
-const createContact = async (req, res) => {
+
+/*const createContact = async (req, res) => {
     //#swagger.tags=['Contacts']
     const contact = {
         firstName: req.body.firstName,
@@ -39,18 +105,43 @@ const createContact = async (req, res) => {
         birthday: req.body.birthday
     };
     const response = await mongodb.getDatabase().db().collection('contacts').insertOne(contact);
-    if (response.acknowledged > 0) {
+    if (response.acknowledged) {
         //res.status(204).send();
         res.status(201).json({ id: response.insertedId });
     }
     else {
         res.status(500).json(response.error || 'Some error occurred while inserting the contact.');
     };
+};*/
+
+const createContact = async (req, res) => {
+    //#swagger.tags=['Contacts']
+    try {
+        const contact = {
+            firstName: req.body.firstName,
+            lastName: req.body.lastName,
+            email: req.body.email,
+            favoriteColor: req.body.favoriteColor,
+            birthday: req.body.birthday
+        };
+        const response = await mongodb.getDatabase().db().collection('contacts').insertOne(contact);
+        if (response.acknowledged) {
+            res.status(201).json({ id: response.insertedId });
+        } else {
+            res.status(500).json(response.error || 'Some error occurred while inserting the contact.');
+        }
+    } catch (err) {
+        res.status(500).json({ message: err.message });
+    }
 };
 
-const updateContact = async (req, res) => {
+/*const updateContact = async (req, res) => {
     //#swagger.tags=['Contacts']
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json('Must use a valid contact id to update a contact.');
+    };
     const contactId = new Objectid(req.params.id);
+    // be aware of updateOne if you only want to update specific fields
     const contact = {
         firstName: req.body.firstName,
         lastName: req.body.lastName,
@@ -59,6 +150,7 @@ const updateContact = async (req, res) => {
         birthday: req.body.birthday
     };
     const response = await mongodb.getDatabase().db().collection('contacts').replaceOne({ _id: contactId }, contact);
+    console.log(response);
     if (response.modifiedCount > 0) {
         res.status(204).send();
     }
@@ -66,18 +158,63 @@ const updateContact = async (req, res) => {
         res.status(500).json(response.error || 'Some error occurred while updateing the contact.');
     };
 
+};*/
+const updateContact = async (req, res) => {
+    //#swagger.tags=['Contacts']
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json('Must use a valid contact id to update a contact.');
+    }
+    try {
+        const contactId = new ObjectId(req.params.id);
+        const contact = {
+            firstName: req.body.firstName,
+            lastName: req.body.lastName,
+            email: req.body.email,
+            favoriteColor: req.body.favoriteColor,
+            birthday: req.body.birthday
+        };
+        const response = await mongodb.getDatabase().db().collection('contacts').replaceOne({ _id: contactId }, contact);
+        if (response.modifiedCount > 0) {
+            res.status(204).send();
+        } else {
+            res.status(500).json(response.error || 'Some error occurred while updating the contact.');
+        }
+    } catch (err) {
+        res.status(500).json({ message: err.message });
+    }
 };
 
-const deleteContact = async (req, res) => {
+/*const deleteContact = async (req, res) => {
     //#swagger.tags=['Contacts']
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json('Must use a valid contact id to delete a contact.');
+    };
     const contactId = new Objectid(req.params.id);
     const response = await mongodb.getDatabase().db().collection('contacts').deleteOne({ _id: contactId });
     if (response.deletedCount > 0) {
         res.status(204).send();
     }
     else {
-        res.status(500).json(response.error || 'Some error occurred while updateing the contact.');
+        res.status(500).json(response.error || 'Some error occurred while deleting the contact.');
     };
+};*/
+
+const deleteContact = async (req, res) => {
+    //#swagger.tags=['Contacts']
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json('Must use a valid contact id to delete a contact.');
+    }
+    try {
+        const contactId = new ObjectId(req.params.id);
+        const response = await mongodb.getDatabase().db().collection('contacts').deleteOne({ _id: contactId });
+        if (response.deletedCount > 0) {
+            res.status(204).send();
+        } else {
+            res.status(500).json(response.error || 'Some error occurred while deleting the contact.');
+        }
+    } catch (err) {
+        res.status(500).json({ message: err.message });
+    }
 };
 
 module.exports = {
